% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{clearReactLog}
\alias{clearReactLog}
\title{Clear the reactive log}
\usage{
clearReactLog()
}
\description{
Convenient way to clear the reactive log without having to start a new
R session or unload Shiny. Normally, the reactive log only grows in size
and never shrinks.
}
\note{
Also clears the reactive graph.
}
