% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{enableReactLog}
\alias{enableReactLog}
\alias{disableReactLog}
\title{Enable or disable reactive event logging}
\usage{
enableReactLog()

disableReactLog()
}
\description{
Convenience functions for \code{options(shiny.reactlog = TRUE)} and
\code{options(shiny.reactlog = NULL)}
}
