% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{getContextNode}
\alias{getContextNode}
\title{Get a reactive context node}
\usage{
getContextNode(x = getCurrentContext(), graph = getReactGraph(),
  invalidated = FALSE)
}
\arguments{
\item{x}{A reactive context, expression, observer. Defaults to the
current reactive context.}

\item{graph}{A reactive graph. Defaults to the reactive graph for the
current Shiny session.}

\item{invalidated}{If \code{TRUE}, get the last invalidated context.}
}
\value{
A reactive context node.
}
\description{
Get a reactive context node
}
\seealso{
Other graph queries: \code{\link{getValueNode}}
}
