% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{getReactGraph}
\alias{getReactGraph}
\title{Get the reactive graph}
\usage{
getReactGraph(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{A Shiny session object or session ID. Defaults to the current
session. If \code{NULL} or no current session, the most recently active session
is used.}
}
\value{
A reactive graph object for the Shiny session.
}
\description{
Builds a reactive dependency graph from the reactive log.
}
\details{
Reactive event logging must be enabled to build graphs:
\code{options(shiny.reactlog = TRUE)} or \code{\link[=enableReactLog]{enableReactLog()}}

To clear the reactive graph, start a new R session or call \code{\link[=clearReactLog]{clearReactLog()}}.
}
\note{
A reactive graph object is created when \code{getReactGraph()} is first
called for a Shiny user session. On subsequent calls to \code{getReactGraph()},
the same graph is retrieved and automatically updated rather than being
completely rebuilt.
}
