% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{getValueNode}
\alias{getValueNode}
\title{Get a reactive value node}
\usage{
getValueNode(x, name = NULL, graph = getReactGraph())
}
\arguments{
\item{x}{A reactive value or reactive values object.}

\item{name}{The name of a value in a reactive values object.}

\item{graph}{A reactive graph. Defaults to the reactive graph for the
current Shiny session.}
}
\value{
A reactive value node.
}
\description{
Get a reactive value node
}
\seealso{
Other graph queries: \code{\link{getContextNode}}
}
