% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{reactlog-package}
\alias{reactlog}
\alias{reactlog-package}
\title{reactlog: Easier Debugging with the Shiny Reactive Log}
\description{
Extends Shiny with new features that make it easier to debug and trace
reactivity, especially in larger apps.
}
\section{Package options}{

\describe{
\item{\code{reactlog.verbose}}{Show detailed information when printing.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/glin/reactlog}
  \item Report bugs at \url{https://github.com/glin/reactlog/issues}
}

}
\author{
\strong{Maintainer}: Greg Lin \email{glin@glin.io}

}
\keyword{internal}
