% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{reactlog_asset_paths}
\alias{reactlog_asset_paths}
\alias{reactlog_add_shiny_resource_paths}
\title{Reactlog asset path information}
\usage{
reactlog_asset_paths()

reactlog_add_shiny_resource_paths()
}
\value{
A \code{list} of \code{list} elements that have
  \code{"prefix"} (string) and \code{"path"} (file path) defined.
}
\description{
To be used internally inside shiny to add resource paths for reactlog.
}
