% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setReactLog.R
\name{reactlog_set}
\alias{reactlog_set}
\title{Set reactlog to be enabled or disabled}
\usage{
reactlog_set(enabled = TRUE)
}
\arguments{
\item{enabled}{whether to enable the reactlog (default) or disable it.}
}
\description{
Before the reactlog can be visualized, it needs to be enabled. It's recommended
to disable the reactlog when deploying a shiny app in production.
}
\seealso{
\code{\link[reactlog]{reactlog_show}}
}
