% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showReactLog.R
\name{reactlog_write}
\alias{reactlog_write}
\title{Write reactlog}
\usage{
reactlog_write(log, file = stdout(), session_token = NULL)
}
\arguments{
\item{log}{produced by shiny to be written}

\item{file}{location of output file. If a \code{NULL} file is given, the
json representation is return.}

\item{session_token}{Session token identifier to be used when subsetting the
complete reactlog}
}
\description{
Write the reactlog to a file.  If a session token is provided, all reactive
interactions will be subsetted to either the global session or the
session provided.
}
