% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer.R
\name{showReactGraph}
\alias{showReactGraph}
\title{Reactive log visualizer}
\usage{
showReactGraph(graph = getReactGraph(), time = TRUE,
  filter = getCurrentContext())
}
\arguments{
\item{graph}{A reactive graph. Defaults to the reactive graph for the
current Shiny session.}

\item{time}{If \code{TRUE}, display the time for each reactive.}

\item{filter}{Filter on a reactive observer or expression and its dependencies.
Defaults to the current reactive context if present.}
}
\description{
Launches the reactive log visualizer with extra features:
\itemize{
\item Interactive filtering on selected nodes
\item Jump to the previous/next flush cycle (i.e. all white nodes)
}
}
\note{
Experimental. Current limitations include:
\itemize{
\item Can only be launched from R like \code{\link[shiny:showReactLog]{shiny::showReactLog()}}
\item Interactive filtering limited to all connected nodes (both dependencies
and dependents)
\item Unable to tell when an observer is destroyed (stays grey forever)
}
}
\examples{
\dontrun{
# show the graph for most recent session
showReactGraph()

observe({
  # show the graph filtered on this observer and its dependencies
  showReactGraph()
})
}
}
\seealso{
\code{\link[shiny:showReactLog]{shiny::showReactLog()}}
}
