% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showReactLog.R
\name{write_reactlog}
\alias{write_reactlog}
\title{Write reactlog}
\usage{
write_reactlog(log, file = stdout(), session_token = NULL)
}
\arguments{
\item{log}{produced by shiny to be written}

\item{file}{location to \code{cat} the file to}

\item{session_token}{Session token identifier to be used when subsetting the complete reactlog}
}
\description{
Write the reactlog to a file.  If a session token is provided, all reactive interactions will be subsetted to either the global session or the session provided.
}
