% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModule.R
\name{reactlog_module_ui}
\alias{reactlog_module_ui}
\alias{reactlog_module_server}
\title{Reactlog Shiny Module}
\usage{
reactlog_module_ui(id = "reactlog", ...)

reactlog_module_server(
  id = "reactlog",
  text_init = "Display Reactlog",
  text_refresh = "Refresh Reactlog",
  width = "100\%",
  height = 600,
  display_immediately = TRUE
)
}
\arguments{
\item{id}{\pkg{shiny} module id to use}

\item{...}{parameters passed to \code{\link[shiny:actionButton]{shiny::actionButton()}}}

\item{text_init}{text to be displayed before the reactlog is displayed.
Clicking the button will cause the reactlog to be loaded.}

\item{text_refresh}{text to be displayed once the reactlog has been displayed.
Clicking the button will cause the reactlog to be reloaded.}

\item{width, height}{HTML attributes to be applied to the reactlog iframe}

\item{display_immediately}{should the reactlog iframe be displayed immediately?}
}
\description{
Displays an iframe of the reactlog in the given application.
}
\details{
State will not be preserved between refreshes.
}
\examples{
if (!require("shiny")) {
  message("`shiny` required to run example")
  return()
}

library(shiny)
# Enable reactlog
reactlog::reactlog_enable()

# Define UI for app that draws a histogram ----
ui <- fluidPage(
  titlePanel("Hello Shiny!"),
  sidebarLayout(
    sidebarPanel(
      sliderInput(inputId = "bins", label = "Number of bins:",
                  min = 1, max = 50, value = 30)
    ),
    mainPanel(plotOutput(outputId = "distPlot"))
  ),
### start ui
  reactlog::reactlog_module_ui(id = "reactlog")
### end ui
)

server <- function(input, output, session) {

  output$distPlot <- renderPlot({
    x    <- faithful$waiting
    bins <- seq(min(x), max(x), length.out = input$bins + 1)
    hist(x, breaks = bins, col = "#75AADB", border = "white",
        xlab = "Waiting time to next eruption (in mins)",
        main = "Histogram of waiting times")
  })

### start server
  reactlog::reactlog_module_server()
### end server

}

if (interactive()) {
  shinyApp(ui = ui, server = server)
}
}
\seealso{
\code{\link[shiny:moduleServer]{shiny::moduleServer()}}
}
